% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{add_join_path}
\alias{add_join_path}
\title{Add a Predefined Join Path to a Query}
\usage{
add_join_path(query_list, ...)
}
\arguments{
\item{query_list}{A xafty query list returned by query()}

\item{...}{List of the join path}
}
\value{
A list
}
\description{
The function gives control over the desired joins in order to bring all projects together into one data.frame. The joins are
read left-to-right. An example can be found under examples on how to structure the list.
}
\examples{
query(projectA = "col1", projectB = "col2", projectC = "col3") |>
add_join_path(path1 = c("projectA", "projectB"), path2 = c("projectA", "projectC"))
# this creates the following joins projectA joins with projectB and projectA joins with projectC
}
