% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_regex_columns_to_validity}
\alias{add_regex_columns_to_validity}
\title{Adds Columns to Validity Matched by Regular Expressions}
\usage{
add_regex_columns_to_validity(check_table, validity_table)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the specified rules in the validity table.}

\item{validity_table}{Data Frame. A validation table that stores the rules that the check table will be checked against.}
}
\description{
The rule ##!!regexcolumns specifies a regular expression as values below the rule that
represents all columns in the check table which match the regular expression. This function
adds these matched columns to the validity.
}
