% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.Date_xafty}
\alias{as.Date_xafty}
\title{Convert Values to Date Types}
\usage{
as.Date_xafty(
  dates,
  date_origin = "1899-12-30",
  force_type = TRUE,
  tryFormats = c("\%Y-\%m-\%d", "\%d.\%m.\%Y", "\%d/\%m/\%Y", "\%Y/\%m/\%d")
)
}
\arguments{
\item{dates}{Character vector of Dates to be Converted}

\item{date_origin}{Character. The date from which numeric dates will be converted into ISO-Date format}

\item{force_type}{Boolean. Whether to force the type conversion even if it introduces NAs during type conversion. If TRUE
the function keeps the column as is.}

\item{tryFormats}{Character vector. Date formats that should be use to try to convert to date}
}
\value{
An equally length date vector, \code{NA} when the value could not be converted to date
}
\description{
Convert Values to Date Types
}
