% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.POSIXct_xafty}
\alias{as.POSIXct_xafty}
\title{Convert Passed Values to POSIXct}
\usage{
as.POSIXct_xafty(datetimes, tz = "")
}
\arguments{
\item{datetimes}{Character vector of date time values to be parsed}

\item{tz}{Timezone for the POSIXct values. Default is UTC}
}
\description{
Convert Passed Values to POSIXct
}
