% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nascent.R
\name{build_dag}
\alias{build_dag}
\title{Build an 'Evaluable' Data Pipeline Object}
\usage{
build_dag(globals, network, frame = "main")
}
\arguments{
\item{globals}{A list returned by dots_to_query()}

\item{network}{A xafty network}

\item{frame}{Used for debugging}
}
\value{
A list
}
\description{
When querying an object, the xafty algorithm recursively iterates through the network and obtains all functions
necessary. Before evaluating all functions, the xafty algorithm creates a dag-object which contains the full
information about dependencies. The object can then be evaluated with function: evaluate_dag
}
