% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_namespace}
\alias{build_namespace}
\title{Main Namespace Generator}
\usage{
build_namespace(link)
}
\arguments{
\item{link}{(list) A link object}
}
\value{
character string of length 1
}
\description{
The namespace builds groups in the execution path, allowing for linear transformations
}
