% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_xafty_list}
\alias{build_xafty_list}
\title{Build a List of Test Results}
\usage{
build_xafty_list(
  check_table,
  validity_table,
  meta_tests_name = "meta_tests",
  check_names = TRUE,
  check_number = TRUE
)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the specified rules in the validity table.}

\item{validity_table}{Data Frame. A validation table that stores the rules that the check table will be checked against.}

\item{meta_tests_name}{Character. Name of the list item that stores all meta_tests. The parameter is there to help avoid naming
conflicts with column names from the check table or validity table.}

\item{check_names}{Boolean. Adds a meta test that checks whether all column names of the validity table are present in
the check table.}

\item{check_number}{Boolean. Adds a meta test that checks whether the number of columns in the check table are equal or
larger than the columns in the validity table.}
}
\value{
A list.
}
\description{
This function is supposed to be an upgrade to the function check_validity(). The xafty_list should make it easy
to efficiently handle all checks and filters for each test within a dashboard context.
}
