% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_xafty_test_table}
\alias{build_xafty_test_table}
\title{Create a Readable Table from a xafty List}
\usage{
build_xafty_test_table(xafty_list)
}
\arguments{
\item{xafty_list}{A list. The list should be the return value from the function build_xafty_list()}
}
\value{
A Dataframe.
}
\description{
The function creates a rather readable output from a xafty list, that can be used in a summary context.
}
