% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_input.R
\name{check_column_exactinput}
\alias{check_column_exactinput}
\title{Check a Table for Exact Value Rules}
\usage{
check_column_exactinput(check_table, validity_table)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the specified rules in the validity table.}

\item{validity_table}{Data Frame. A table that stores the rules by which the check table is compared to.}
}
\description{
Check a Table for Exact Value Rules
}
