% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_number.R
\name{check_column_number}
\alias{check_column_number}
\title{Check for Correct Number of Columns}
\usage{
check_column_number(check_table, validity_table)
}
\description{
Check for Correct Number of Columns
}
