% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_types.R
\name{check_column_types}
\alias{check_column_types}
\title{Check Column Classes}
\usage{
check_column_types(check_table, validity_table, simply = FALSE)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the specified rules in the validity table.}

\item{validity_table}{Data Frame. A table that stores the rules by which the check table is compared to.}

\item{simply}{Boolean. Changes the return value of the function to a single logical vector of length 1.}
}
\description{
Check Column Classes
}
