% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_result_row}
\alias{create_result_row}
\title{Row Template for Result Data Frame}
\usage{
create_result_row(
  check_name,
  default_result = TRUE,
  default_message = "NOT CHECKED!",
  default_columns = NA
)
}
\arguments{
\item{check_name}{Character string specifying the name of the check to be performed.}

\item{default_result}{Logical value representing the default result of the check. Default is \code{TRUE}.}

\item{default_message}{Character string representing the default message for the check result. Default is \code{"NOT CHECKED!"}.}

\item{default_columns}{Default columns to be included in the result data frame. Default is \code{NA}.}
}
\description{
Row Template for Result Data Frame
}
