% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nascent.R
\name{evaluate_dag}
\alias{evaluate_dag}
\title{Evaluate a Network Pipeline}
\usage{
evaluate_dag(dag)
}
\arguments{
\item{dag}{List-Object created by [xafty::build_dag]}
}
\value{
A data.frame
}
\description{
Evaluate the pipeline created via [xafty::build_dag]
}
