% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_column_empty.R
\name{filter_column_empty}
\alias{filter_column_empty}
\title{Filter Values in Column That are NA}
\usage{
filter_column_empty(check_table, filter_column)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the specified rules in the validity table.}

\item{filter_column}{Character. The column which will be checked. The column must be present in both the check table
as well as the validity table}
}
\value{
A logical vector.
}
\description{
Filter Values in Column That are NA
}
