% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_names.R
\name{filter_column_names}
\alias{filter_column_names}
\title{Filter for Column Names}
\usage{
filter_column_names(check_table, validity_table)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the information in the validity table}

\item{validity_table}{Data Frame. A table that stores the column names in the first row}
}
\description{
The function filters for column names in two ways:
* Column names in the check table that are expected by the validity table
* Expected columns by the validity table that are not in the check table.

This function also tries to implement column names that are stored in the validity table as regular expressions.
}
