% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_column_type.R
\name{filter_column_type}
\alias{filter_column_type}
\title{Filter a Column by its Data Type given the Validity}
\usage{
filter_column_type(
  check_table,
  validity_table,
  filter_column,
  date_origin = "1899-12-30"
)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the information in the validity table}

\item{validity_table}{Data Frame. A table that stores the column names in the first row}

\item{filter_column}{Character. The column which will be checked. The column must be present in both the check table
as well as the validity table}

\item{date_origin}{Character. The date from which numeric dates will be conversed into ISO-Date format}
}
\value{
A Boolean vector. TRUE when the value is in line with the specified data type; FALSE when it is not
}
\description{
Filter a Column by its Data Type given the Validity
}
