% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{from}
\alias{from}
\title{Specify the Main Project for a Query}
\usage{
from(query_list, project)
}
\arguments{
\item{query_list}{A query object returned from query()}

\item{project}{Name of the project}
}
\value{
A xafty Query
}
\description{
Specify the Main Project for a Query
}
