% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_xafty_list_items}
\alias{get_xafty_list_items}
\title{Gets the Desired Items from a Column}
\usage{
get_xafty_list_items(xafty_list, column, item)
}
\arguments{
\item{xafty_list}{A List Created by build_xafty_list()}

\item{column}{The column from which the items should be taken?}

\item{item}{The desired item from the xafty_list}
}
\description{
Gets the Desired Items from a Column
}
