% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{init_network}
\alias{init_network}
\title{Initialize a New xafty Network}
\usage{
init_network()
}
\value{
A 'xafty_network' environment
}
\description{
Use this function to initialize a new xafty network.
}
\examples{
# Initialize the project
new_network <- init_network()

# Add a new project
new_network$add_project("project1")

}
