% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{init_project}
\alias{init_project}
\title{Initialize a New xafty Project}
\usage{
init_project(project = NULL, path = paste0(getwd(), "/R"))
}
\arguments{
\item{path}{Path to where the project should be initialized}

\item{name}{Name of the project. The project name needs to be a syntactically valid character vector of length 1.}
}
\value{
A 'xafty_network' list of environments
}
\description{
The function creates a new sub folder to the specified path where all xafty objects will be stored
}
\examples{
\dontrun{
init_project(name = "my_valid_project_name", path = getwd())
}
}
