% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date_xafty}
\alias{is.Date_xafty}
\title{Check if Passed Values can be Parsed as Date}
\usage{
is.Date_xafty(dates, date_origin = "1899-12-30")
}
\arguments{
\item{dates}{Character vector of Dates to be Parsed}

\item{date_origin}{Character. The date from which numeric dates will be conversed into ISO-Date format}
}
\value{
An equally length boolean vector whether the value can be parsed as a Date given the specified formats and origin
}
\description{
Check if Passed Values can be Parsed as Date
}
