% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.numeric_xafty}
\alias{is.numeric_xafty}
\title{Check if Passed Values can be Parsed as Numeric}
\usage{
is.numeric_xafty(numbers)
}
\arguments{
\item{numbers}{Character vector of Numbers to be Parsed}
}
\value{
An equally length Boolean vector whether the values can be parsed as numbers
}
\description{
Check if Passed Values can be Parsed as Numeric
}
