% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.POSIXct_xafty}
\alias{is.POSIXct_xafty}
\title{Check if Passed Values can be Parsed as POSIXct}
\usage{
is.POSIXct_xafty(datetimes)
}
\arguments{
\item{datetimes}{Character vector of date time values to be parsed}
}
\value{
An equally length Boolean vector whether the values can be parsed as POSIXct
}
\description{
Check if Passed Values can be Parsed as POSIXct
}
