% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{merge_networks}
\alias{merge_networks}
\title{Merge networks into one}
\usage{
merge_networks(name, ...)
}
\arguments{
\item{name}{Character vector of length 1. Name of the merged network}

\item{...}{Networks that should be merged}
}
\value{
A 'xafty_network' environment
}
\description{
The function allows the merging of any numbers of networks into one network. The return value is therefore always one network.
}
