% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nascent.R
\name{nascent}
\alias{nascent}
\title{Retrieve Data from a xafty network}
\usage{
nascent(network, xafty_list)
}
\arguments{
\item{network}{A xafty network}

\item{xafty_list}{Xafty link object created with pull_link}
}
\description{
Retrieve Data from a xafty network
}
