% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_columns_in_validity}
\alias{obtain_columns_in_validity}
\title{Checks Validity Table for Presence of Xafty Syntax}
\usage{
obtain_columns_in_validity(validity_table, xafty_syntax)
}
\arguments{
\item{validity_table}{A validity table that will be checked for xafty rules}

\item{xafty_syntax}{The xafty rules which presence will be checked in the validity table}
}
\value{
Returns a named vector that has the column name as value and the rule as the name. This named vector is also
referred to as "xafty pair".
}
\description{
Checks Validity Table for Presence of Xafty Syntax
}
