% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_invalid_columns}
\alias{obtain_invalid_columns}
\title{Obtain the Column Names Separately}
\usage{
obtain_invalid_columns(column_string)
}
\arguments{
\item{column_string}{Column names as a single string that have been separated by ", "}
}
\value{
The column names in a vector
}
\description{
Obtain the Column Names Separately
}
