% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_values_in_validity}
\alias{obtain_values_in_validity}
\title{Get Corresponding Values for Xafty Rules}
\usage{
obtain_values_in_validity(validity_table, xafty_pair)
}
\arguments{
\item{validity_table}{A validity table that will be checked for xafty rules}

\item{xafty_pair}{A named vector of length 1 that is named with the xafty rule and has the value of a column name}
}
\description{
Get Corresponding Values for Xafty Rules
}
