% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.xafty_network}
\alias{print.xafty_network}
\title{Print a xafty Network}
\usage{
\method{print}{xafty_network}(x, ...)
}
\arguments{
\item{x}{an object with the class "xafty_network"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 method to print a xafty network.
}
