% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.xafty_project}
\alias{print.xafty_project}
\title{Print a xafty Project}
\usage{
\method{print}{xafty_project}(x, ...)
}
\arguments{
\item{x}{an object with the class "xafty_project"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 method to print a xafty network. Variables can be organized into groups;
grouped variables are displayed separately while maintaining layer hierarchy within each group.
Ungrouped variables are displayed using the traditional layer-based organization.
}
