% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Build a xafty Query}
\usage{
query(...)
}
\arguments{
\item{...}{Query value. See examples for creating a simple query}
}
\description{
A xafty query will be passed to the nascent function in order to retrieve data from the network
}
\examples{
query(project_name1 = c("col1", "col2"), project_name2 = c("colA"))
}
