% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_check_and_validity}
\alias{read_check_and_validity}
\title{Reads in Example Data for Test Purposes}
\usage{
read_check_and_validity(
  path_check = "inst/extdata/example_data.xlsx",
  path_validity = "inst/extdata/example_validity.csv",
  file_ending = "xlsx"
)
}
\arguments{
\item{path_check}{Path to the table that will be checked}

\item{path_validity}{Path to the validity table that stores the accompanying xafty rules}

\item{file_ending}{Character. Valid input: "xlsx", "csv_comma", "csv_semicolon". You need to make sure that the file ending
matches the actual file.}
}
\description{
Reads in Example Data for Test Purposes
}
