% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleset.R
\name{set_project_print_order}
\alias{set_project_print_order}
\title{Change Print Order of Projects}
\usage{
set_project_print_order(projects, order = NULL, network)
}
\arguments{
\item{projects}{Character of project names of which the order should be changed}

\item{order}{Numeric vector with equal length of parameter projects to set the corresponding print order}

\item{network}{A xafty network containing the projects}
}
\value{
A xafty network invisibly
}
\description{
The function allows to manually set the order for each project in the network. It is also possible to set the order
for several projects at once by providing an equally length order vector.
}
\examples{
xafty_network <- init_network("change_order", projects = c("Project2", "Project1"))
print(xafty_network) # Prints Project2 first and Project1 second
set_project_print_order(projects = c("Project2", "Project1"),
                        order = c(2, 1),
                        network = xafty_network)
print(xafty_network)
}
