% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_script.R
\name{to_script}
\alias{to_script}
\title{Write a Pipeline to a Script}
\usage{
to_script(dag, file = NULL)
}
\arguments{
\item{dag}{A directed acyclic graph object obtained by build_dag()}

\item{file}{An R-script file to write to}
}
\value{
TRUE (invisibly)
}
\description{
This function is used in conjunction with build_dag() to write the created pipeline as a linear script. This is useful
for debugging purposes to see on which step the pipeline fails. The function creates a single file which can be run from
top to bottom as it would be evaluated in nascent()
}
