% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Filter a Network Pipeline}
\usage{
where(query_list, ...)
}
\arguments{
\item{query_list}{A xafty query obtained via [`xafty::query`]}

\item{...}{Filter conditions, e.g. `col1 >= 1```}
}
\value{
A xafty query with a where filter applied
}
\description{
Filter a Network Pipeline
}
