% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{with}
\alias{with}
\title{Add a State to a xafty Query}
\usage{
with(query_list, ...)
}
\arguments{
\item{query_list}{A query list created by [xafty::query]}

\item{...}{Declaration of the states e.g. state_name = TRUE}
}
\value{
A list state_query which is a query bundled with a state
}
\description{
The state of a xafty query is passed to arguments that were declared as xafty states
by passing a single character vector wrapped into {curley_braces} into the argument
}
