% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xafty_rules.R
\docType{data}
\name{xafty_rules_table}
\alias{xafty_rules_table}
\title{Table of All Xafty Rules}
\format{
A data frame with 6 columns:
\describe{
  \item{syntax}{A character vector with xafty rules.}
  \item{type}{A character vector of rule types.}
  \item{description}{A character vector of rule description in prosa.}
  \item{check_function}{A list of functions for check purposes.}
  \item{filter_function}{A list of functions to filter values.}
  \item{change_type_function}{A list of functions change data to appropriate types.}
}
}
\source{
The data is curated by xafty developers and grows as new rules are added.
}
\usage{
xafty_rules_table
}
\description{
A simple table containing the xafty rules, their type, a description, and useful functions
}
\keyword{datasets}
