% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_names.R
\name{add_regex_columns_to_validity}
\alias{add_regex_columns_to_validity}
\title{Adds Columns to Validity Matched by Regular Expressions}
\usage{
add_regex_columns_to_validity(check_table, validity_table, multiple = "remove")
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the specified rules in the validity table.}

\item{validity_table}{Data Frame. A validation table that stores the rules that the check table will be checked against.}

\item{multiple}{Character.
The following Parameters are possible:
\itemize{
 \item "remove": The default. Removes all columns that have been duplicated through multiple matches
 \item "keep": Keeps all columns that have been duplicated through multiple matches
}}
}
\value{
A Data Frame. The return value is the new validity table with column names added from the check table matched
by the regular expressions. Each column matched in this way also inherits the same set of rules specified
in the same column as the ##!!regexcolumns rule.
}
\description{
The rule ##!!regexcolumns specifies a regular expression as values below the rule that
represents all columns in the check table which match the regular expression. This function
adds these matched columns to the validity.
}
