% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_classes.R
\name{align_column_classes}
\alias{align_column_classes}
\title{Align Column Classes with Validity Table}
\usage{
align_column_classes(check_table, validity_table)
}
\description{
This function changes the column classes according to the data types of the validity table
}
