% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_names.R
\name{check_column_names}
\alias{check_column_names}
\title{Check for Column Names}
\usage{
check_column_names(check_table, validity_table, check_order = FALSE)
}
\description{
Check for Column Names
}
