% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_empty.R
\name{check_column_notempty}
\alias{check_column_notempty}
\title{Checks for Empty Entries}
\usage{
check_column_notempty(check_table, validity_table)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the specified rules in the validity table.}

\item{validity_table}{Data Frame. A table that stores the rules by which the check table is compared to.}
}
\description{
Checks for Empty Entries
}
