% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_check_validity.R
\name{check_validity}
\alias{check_validity}
\title{Main Function to Check Validity of a Table}
\usage{
check_validity(
  path_table,
  path_validity,
  file_ending = "xlsx",
  check_number = TRUE,
  check_names = TRUE,
  check_classes = TRUE
)
}
\arguments{
\item{path_table}{Character. A file path to the table will be checked for validity}

\item{file_ending}{Character. Valid input: "xlsx", "csv_comma", "csv_semicolon". You need to make sure that the file ending
matches the actual file.}

\item{check_number}{Boolean. Should the number of columns be checked?}

\item{check_classes}{Boolean. Should the column's classes be checked?}

\item{path_valid}{Character. A file path to a special csv file that contains the valid syntax}

\item{check_name}{Boolean. Should the column names be checked?}
}
\description{
Main Function to Check Validity of a Table
}
