% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_check_validity.R
\name{check_validity}
\alias{check_validity}
\title{Main Function to Check a Table for its Validity}
\usage{
check_validity(
  check_table,
  validity_table,
  column_number = "equal",
  column_names = "presence",
  column_types = TRUE,
  values_notempty = TRUE,
  values_exact = TRUE,
  values_pattern = TRUE
)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the specified rules in the validity table.}

\item{validity_table}{Data Frame. A validation table that stores the rules that the check table will be checked against.}

\item{column_number}{Character or Boolean. Should the count of columns be checked?
Following Parameters are allowed:
\itemize{
 \item "equal": If number of columns in both tables should be equal
 \item "larger": If number of columns in both tables should be equal or larger
 \item FALSE: If column number should not be checked
}}

\item{column_names}{Character or Boolean. Should the column names be checked
Following parameters are allowed:
\itemize{
 \item "presence": If the simple presence of each column suffices
 \item "order": If the columns should be present as well as in the specified order
 \item FALSE: If column names should not be checked
}}

\item{column_types}{Boolean. Should the column's data types be checked?}

\item{values_notempty}{Boolean. Should columns be checked for NA entries?}

\item{values_exact}{Boolean. Should columns be checked for exact value rules?}

\item{values_pattern}{Boolean. Should columns be checked for pattern value rules?}
}
\value{
A data frame that stores the check results.
The data frame will contain the following columns:
\itemize{
 \item Check: Contains the names of the rule
 \item Check_Result: Boolean values representing the result of the check. Check passed: TRUE, check failed: FALSE
 \item Message: Detailed information about the check result
 \item Columns: Name of the column(s) where the respective rule was broken
}
}
\description{
The function checks a table (the check table) for its validity using a table (the validity table)
that contains the xafty rules to specify the expected column names and values.

\strong{!! Important:} You will need to make sure that empty values are correctly identified as NA.
}
