% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_check_validity.R
\name{check_validity}
\alias{check_validity}
\title{Main Function to Check a Table for its Validity}
\usage{
check_validity(
  check_table,
  validity_table,
  column_number = TRUE,
  column_names = "presence",
  column_classes = TRUE,
  values_notempty = TRUE,
  values_exact = TRUE,
  values_pattern = "##!!strictpattern"
)
}
\arguments{
\item{check_table}{Data Frame. The table that will be checked against the information in the validity table
Important: You will need to make sure that empty values are correctly identified with NA.}

\item{validity_table}{Data Frame. A table that stores the rules by which the check table is compared to.}

\item{values_notempty}{Boolean. Should columns be checked for NA entries?}

\item{values_exact}{Boolean. Should columns be checked for the exact value rules?}

\item{values_pattern}{Boolean. Should columns be checked for the pattern value rules?}

\item{col_number}{Boolean. Should the number of columns be checked?}

\item{col_names}{´Character. Following parameters are allowed:
\itemize{
 \item "presence": If the simple presence of each column suffices.
 \item "order": If the columns should be present as well as in the specified order within the validity table.
 \item FALSE: If column names should not be checked
}}

\item{col_classes}{Boolean. Should the column's classes be checked?}
}
\value{
Data Frame. A table that stores the check results.
}
\description{
The function checks a table (the check table) for its validity using a table (the validity table)
that is crafted with xafty rules to specify the expected column names and column values.
}
