% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{init_network}
\alias{init_network}
\title{Initialize a New xafty Network}
\usage{
init_network(name, projects = NULL)
}
\arguments{
\item{name}{Character vector of length 1. A valid name for the network}

\item{projects}{Character vector of project names that should be added to the network}
}
\value{
A 'xafty_network' environment
}
\description{
Use this function to initialize a new xafty network.
}
\examples{
{
# Initialize the network
new_network <- init_network(name = "network_1")

# Add a new project
new_network$add_project("project1")
}
}
