% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nascent.R
\name{nascent}
\alias{nascent}
\title{Retrieve Data from a xafty network}
\usage{
nascent(network, ...)
}
\arguments{
\item{network}{A xafty network}

\item{...}{List.}
}
\description{
Retrieve Data from a xafty network
}
