% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register a Function to a xafty Network}
\usage{
register(
  fun,
  link_type = c("get", "add", "join"),
  module = "link",
  network_env = NULL,
  project = NULL
)
}
\description{
Register a Function to a xafty Network
}
