% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register a Function in a Network}
\usage{
register(quosure, project, network, link_type, ...)
}
\arguments{
\item{quosure}{A function call enquoted with rlang::enquo}

\item{project}{The project name of the project within the network where the function should be registered.}

\item{network}{A xafty network.}

\item{link_type}{The link_type name of the rulset. Currently "link" and "object" are supported}

\item{...}{Configurations and advanced use when registering a new link}
}
\value{
A xafty network (invisibly)
}
\description{
The function is the workhorse behind all register methods of a network.
}
