% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register a Function in a Network}
\usage{
register(quosure, project, network, module, ...)
}
\arguments{
\item{quosure}{A function call enquoted with rlang::enquo}

\item{project}{The project name of the project within the network where the function should be registered.}

\item{network}{A xafty network.}

\item{module}{The module name of the rulset. Currently only "link" is supported}

\item{...}{Configurations of the register. Following parameters are allowed
* added_columns : The parameter takes in a character vector of column names that are added to the data set.}
}
\value{
A xafty network (invisibly)
}
\description{
The function is the workhorse behind all register methods of a network.
}
